requirements.txt
beautifulsoup4==4.11.1
dbTable==0.0.4
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
Pillow==9.1.1
plotly==5.2.2
psycopg2==2.9.3
psycopg2_binary==2.9.3
python_bcrypt==0.3.2
requests==2.26.0
SQLAlchemy==1.4.22
streamlit==1.10.0
streamlit_multipage==0.0.18
streamlit_option_menu==0.3.2
yagmail==0.15.277
~treamlit==1.9.0
